End Screen, HighScore, Death for gate and ghost
Ghost can now end the game when killed by a zombie, to accomplish this I put a bool in the ghost during possession that blocks the destroy code from running.

End screen shows the built in gamemaker highscores. This does not save to files, and looks kind of gross. TBH it could be replaced with a single line showing the score, but I like the placeholder?

The gate on destroy is just PlayerDeath();

I had to create a reset button, and to do this I copied the old buttons. These were programmed in the Visual langauge instead of GML, So I converted them for simplicity on my end.

End screen routes properly to different maps.

oController made persisitent!! I then use reset_game(); to start the game on the end screen. This matters to reset the leaderboard, but otherwise not.